[MID-21] add placeholder metadata to pages
