+++ src/app/admin/layout.js
@@ -1,6 +1,6 @@
 export const metadata = {
   title: 'Admin Panel',
-  description: 'Generated by create next app',
+  description: 'Admin panel for Mayak',
 };
 
 export default function AdminLayout({ children }) {
