+++ src/app/layout.js
@@ -9,8 +9,10 @@ import { QueryContext } from './queryContext';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata = {
-  title: 'Mayak!',
-  description: 'Generated by create next app',
+  title: {
+    template: '%s | Маяк',
+    default: 'Маяк',
+  },
 };
 
 export default function RootLayout({ children }) {
@@ -18,7 +20,9 @@ export default function RootLayout({ children }) {
     <html lang="en">
       <body className={inter.className}>
         <Header />
-        <QueryContext>{children}</QueryContext>
+        <QueryContext>
+          <main>{children}</main>
+        </QueryContext>
         <Footer />
       </body>
     </html>
