+++ src/app/page.js
@@ -1,5 +1,9 @@
 import React from 'react';
 
-export default function Home() {
+export const metadata = {
+  description: 'Опис головної сторінки',
+};
+
+export default function Page() {
   return <div>Main</div>;
 }
